pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "daisy-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy-ai>=2.6",
]
[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/daisy"]
sources = ["src"]
[tool.hatch.build.targets.sdist]
include = ["src/daisy/py.typed"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
